requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
Flask-Migrate==4.0.5
python-dateutil==2.8.2
Flask-SocketIO==5.3.6
eventlet==0.33.3
bidict==0.22.1
APScheduler==3.10.1
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
pytest-flask==1.3.0
